[refactor] Merge ``_compare_eq_cls`` into ``_compare_any.py`` with the dispatcher
The two functions mutually recurse: ``_compare_eq_any`` dispatches to
``_compare_eq_cls`` for dataclasses/attrs/namedtuples, and
``_compare_eq_cls`` calls ``_compare_eq_any`` to drill into differing
attributes. Splitting them across modules forces a circular import that
no longer resolves now that the guards have moved out of ``util.py``,
so they live together as one logical unit. ``_compare_cls.py`` is
removed.
